internal/keyspan: consolidate datadriven iterator ops
Previously, a few different datadriven test repeated logic for driving a
FragmentIterator through datadriven commands. This commit refactors them to
share code for driving these datadriven commands.
